sistema_gestion_agricola/requirements.txt
Flask==2.2.5
Flask-Caching==1.11.1
python-dotenv==1.0.1
requests==2.31.0
pdfkit==1.0.0
Werkzeug==2.2.3
gunicorn==20.1.0
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.9
WeasyPrint==60.1
